backend/requirements.txt
# Web Framework
fastapi==0.115.1
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
watchfiles==0.21.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
pgvector>=0.3.0

# Data Validation
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Job Queue
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# AI/ML - LLM Providers (pinned so pip resolves fast)
google-genai
anthropic==0.39.0
openai==1.12.0

# MCP - Model Context Protocol (server + client)
mcp==1.2.0

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Telegram
python-telegram-bot>=21.0

# Utilities
python-dotenv==1.0.0
pydantic-extra-types>=2.1.0
bleach>=6.1.0                  # XSS sanitization

# Logging
python-json-logger==2.0.7

# Resilience — retry + circuit breaker
tenacity==8.5.0
pybreaker==1.2.0

# Conversation state machine
transitions==0.9.2

# Observability — OpenTelemetry
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-exporter-otlp-proto-http>=1.24.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-sqlalchemy>=0.44b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov>=4.1.0               # Test coverage
aiosqlite>=0.19.0               # SQLite async for testing

# Date/Time utilities
python-dateutil==2.8.2
pytz==2023.3
